+++ internal/app/interfaces.go
@@ -13,7 +13,7 @@ type UserService interface {
 }
 
 type FarmerService interface {
-	CreateFarmer(name string, chickenBalance int32, cashBalance int32) (*domain.Farmer, error)
+	CreateFarmer(name string, chickenBalance float64, cashBalance int32) (*domain.Farmer, error)
 	GetFarmerByName(name string) (*domain.Farmer, error)
 	GetFarmers() ([]domain.Farmer, error)
 	DeleteFarmerByID(ID uuid.UUID) error
